Move logic from config.go
