+++ config.go
@@ -15,18 +15,10 @@
 package activemq
 
 import (
-	"bytes"
 	"context"
-	"crypto/tls"
-	"crypto/x509"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"os"
 	"time"
 
 	sdk "github.com/conduitio/conduit-connector-sdk"
-	"github.com/go-stomp/stomp/v3"
 )
 
 //go:generate paramgen -output=paramgen_src.go SourceConfig
@@ -106,122 +98,4 @@ type DestinationConfig struct {
 	// DestinationHeader maps to the "destination" header in the STOMP SEND
 	// frame. Useful when using ANYCAST.
 	DestinationHeader string `json:"destinationHeader"`
-}
-
-type Position struct {
-	MessageID   string `json:"message_id"`
-	Destination string `json:"destination"`
-}
-
-func parseSDKPosition(sdkPos sdk.Position) (Position, error) {
-	decoder := json.NewDecoder(bytes.NewBuffer(sdkPos))
-	decoder.DisallowUnknownFields()
-
-	var p Position
-	if err := decoder.Decode(&p); err != nil {
-		return p, fmt.Errorf("failed to decode position: %w", err)
-	}
-	return p, nil
-}
-
-func (p Position) ToSdkPosition() sdk.Position {
-	bs, err := json.Marshal(p)
-	if err != nil {
-		// this should never happen
-		panic(err)
-	}
-
-	return sdk.Position(bs)
-}
-
-// metadataFromMsg extracts all the present headers from a stomp.Message into
-// sdk.Metadata.
-func metadataFromMsg(msg *stomp.Message) sdk.Metadata {
-	metadata := make(sdk.Metadata)
-
-	for i := range msg.Header.Len() {
-		k, v := msg.Header.GetAt(i)
-		metadata[k] = v
-	}
-
-	return metadata
-}
-
-func connect(ctx context.Context, config Config) (*stomp.Conn, error) {
-	connOpts := []func(*stomp.Conn) error{
-		stomp.ConnOpt.Login(config.User, config.Password),
-		stomp.ConnOpt.HeartBeat(config.SendTimeoutHeartbeat, config.RecvTimeoutHeartbeat),
-	}
-
-	if !config.TLS.Enabled {
-		conn, err := stomp.Dial("tcp", config.URL, connOpts...)
-		if err != nil {
-			return nil, fmt.Errorf("failed to connect to ActiveMQ: %w", err)
-		}
-		sdk.Logger(ctx).Debug().Msg("opened connection to ActiveMQ")
-
-		return conn, nil
-	}
-
-	sdk.Logger(ctx).Debug().Msg("using TLS to connect to ActiveMQ")
-
-	cert, err := tls.LoadX509KeyPair(config.TLS.ClientCertPath, config.TLS.ClientKeyPath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load client key pair: %w", err)
-	}
-	sdk.Logger(ctx).Debug().Msg("loaded client key pair")
-
-	caCert, err := os.ReadFile(config.TLS.CaCertPath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load CA cert: %w", err)
-	}
-	sdk.Logger(ctx).Debug().Msg("loaded CA cert")
-
-	caCertPool := x509.NewCertPool()
-	caCertPool.AppendCertsFromPEM(caCert)
-
-	tlsConfig := &tls.Config{
-		MinVersion: tls.VersionTLS12,
-		MaxVersion: tls.VersionTLS13,
-
-		Certificates: []tls.Certificate{cert},
-		RootCAs:      caCertPool,
-
-		InsecureSkipVerify: config.TLS.InsecureSkipVerify, // #nosec G402
-	}
-
-	netConn, err := tls.Dial("tcp", config.URL, tlsConfig)
-	if err != nil {
-		panic(err)
-	}
-	sdk.Logger(ctx).Debug().Msg("TLS connection established")
-
-	conn, err := stomp.Connect(netConn, connOpts...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to connect to ActiveMQ: %w", err)
-	}
-	sdk.Logger(ctx).Debug().Msg("STOMP connection using tls established")
-
-	return conn, nil
-}
-
-func teardown(ctx context.Context, subs *stomp.Subscription, conn *stomp.Conn, label string) error {
-	if subs != nil {
-		err := subs.Unsubscribe()
-		if errors.Is(err, stomp.ErrCompletedSubscription) {
-			sdk.Logger(ctx).Debug().Msg("subscription already unsubscribed")
-		} else if err != nil {
-			return fmt.Errorf("failed to unsubscribe: %w", err)
-		}
-	}
-
-	if conn != nil {
-		if err := conn.Disconnect(); err != nil {
-			return fmt.Errorf("failed to disconnect from ActiveMQ: %w", err)
-		}
-	}
-
-	sdk.Logger(ctx).Debug().Msgf("teardown for %s complete", label)
-
-	return nil
 }
